Commented code, added args
Commented all/most of the code, added start_time and end_time args to the parser so you can now start and stop a recording when you want to. Good for downloading just matches etc.

Moved the FFMpeg combine videos function to the downloader.

Implemented the base of the part downloader. I think I will need to revert back to segment_number instead of current_time.

Moved the writing of the data to it's own function instead of the download_playlist function.